@@ -1,83 +1,117 @@
 {
   "name": "jiti",
-  "version": "1.20.0",
+  "version": "2.0.0-rc.1",
   "description": "Runtime typescript and ESM support for Node.js",
   "repository": "unjs/jiti",
   "license": "MIT",
-  "main": "./lib/index.js",
-  "types": "dist/jiti.d.ts",
-  "bin": "bin/jiti.js",
+  "type": "module",
+  "exports": {
+    ".": {
+      "import": {
+        "types": "./lib/jiti.d.mts",
+        "default": "./lib/jiti.mjs"
+      },
+      "require": {
+        "types": "./lib/jiti.d.cts",
+        "default": "./lib/jiti.cjs"
+      }
+    },
+    "./register": {
+      "types": "./lib/jiti-register.d.mts",
+      "import": "./lib/jiti-register.mjs"
+    },
+    "./native": {
+      "types": "./lib/jiti.d.mts",
+      "import": "./lib/jiti-native.mjs"
+    },
+    "./package.json": "./package.json"
+  },
+  "main": "./lib/jiti.cjs",
+  "module": "./lib/jiti.mjs",
+  "types": "./lib/jiti.d.cts",
+  "bin": {
+    "jiti": "./lib/jiti-cli.mjs"
+  },
   "files": [
     "lib",
     "dist",
-    "register.js"
+    "register.cjs"
   ],
   "scripts": {
     "build": "pnpm clean && NODE_ENV=production pnpm webpack",
     "clean": "rm -rf dist",
     "dev": "pnpm clean && pnpm webpack --watch",
-    "jiti": "JITI_DEBUG=1 JITI_CACHE=false JITI_REQUIRE_CACHE=false ./bin/jiti.js",
-    "jiti:legacy": "JITI_DEBUG=1 npx node@12 ./bin/jiti.js",
-    "lint": "eslint --ext .ts,.js . && prettier -c src lib test stubs",
-    "lint:fix": "eslint --fix --ext .ts,.js . && prettier -w src lib test stubs",
-    "release": "pnpm build && pnpm test && changelogen --release --push && npm publish",
-    "test": "pnpm lint && vitest run --coverage && pnpm test:bun",
-    "test:bun": "bun --bun test test/bun"
+    "jiti": "JITI_DEBUG=1 JITI_JSX=1 lib/jiti-cli.mjs",
+    "lint": "eslint . && prettier -c src lib test stubs",
+    "lint:fix": "eslint --fix . && prettier -w src lib test stubs",
+    "prepack": "pnpm build",
+    "release": "pnpm build && pnpm test && changelogen --release --prerelease --push --publish --publishTag 2x",
+    "test": "pnpm lint && pnpm test:types && vitest run --coverage && pnpm test:node-register && pnpm test:bun && pnpm test:native",
+    "test:bun": "bun --bun test test/bun",
+    "test:native": "pnpm test:native:bun && pnpm test:native:node && pnpm test:native:deno",
+    "test:native:bun": "bun --bun test test/native/bun.test.ts",
+    "test:native:deno": "deno test -A test/native/deno.test.ts",
+    "test:native:node": "node --test --experimental-strip-types test/native/node.test.ts",
+    "test:node-register": "JITI_JSX=1 node --test test/node-register.test.mjs",
+    "test:types": "tsc --noEmit"
   },
   "devDependencies": {
-    "@babel/core": "^7.22.15",
-    "@babel/helper-plugin-utils": "^7.22.5",
-    "@babel/plugin-proposal-decorators": "^7.22.15",
+    "@babel/core": "^7.25.2",
+    "@babel/helper-module-imports": "^7.24.7",
+    "@babel/helper-module-transforms": "^7.25.2",
+    "@babel/helper-plugin-utils": "^7.24.8",
+    "@babel/helper-simple-access": "^7.24.7",
+    "@babel/plugin-proposal-decorators": "^7.24.7",
     "@babel/plugin-syntax-class-properties": "^7.12.13",
-    "@babel/plugin-syntax-import-assertions": "^7.22.5",
-    "@babel/plugin-transform-export-namespace-from": "^7.18.9",
-    "@babel/plugin-transform-modules-commonjs": "^7.22.15",
-    "@babel/plugin-transform-nullish-coalescing-operator": "^7.18.6",
-    "@babel/plugin-transform-optional-chaining": "^7.21.0",
-    "@babel/plugin-transform-typescript": "^7.22.15",
-    "@babel/preset-typescript": "^7.22.15",
-    "@babel/template": "^7.22.15",
-    "@babel/traverse": "^7.22.15",
-    "@babel/types": "^7.22.15",
-    "@types/babel__core": "^7.20.1",
-    "@types/babel__template": "^7.4.1",
-    "@types/babel__traverse": "^7.20.1",
-    "@types/node": "^20.5.9",
-    "@types/object-hash": "^3.0.4",
-    "@types/resolve": "^1.20.2",
-    "@types/semver": "^7.5.1",
-    "@vitest/coverage-v8": "^0.34.3",
-    "acorn": "^8.10.0",
-    "babel-plugin-dynamic-import-node": "^2.3.3",
+    "@babel/plugin-syntax-import-assertions": "^7.25.6",
+    "@babel/plugin-syntax-jsx": "^7.24.7",
+    "@babel/plugin-transform-export-namespace-from": "^7.24.7",
+    "@babel/plugin-transform-react-jsx": "^7.25.2",
+    "@babel/plugin-transform-typescript": "^7.25.2",
+    "@babel/preset-typescript": "^7.24.7",
+    "@babel/template": "^7.25.0",
+    "@babel/traverse": "^7.25.6",
+    "@babel/types": "^7.25.6",
+    "@types/babel__core": "^7.20.5",
+    "@types/babel__helper-module-imports": "^7.18.3",
+    "@types/babel__helper-plugin-utils": "^7.10.3",
+    "@types/babel__template": "^7.4.4",
+    "@types/babel__traverse": "^7.20.6",
+    "@types/node": "^22.5.5",
+    "@vitest/coverage-v8": "^2.1.1",
+    "acorn": "^8.12.1",
     "babel-plugin-parameter-decorator": "^1.0.16",
-    "changelogen": "^0.5.5",
-    "config": "^3.3.9",
-    "create-require": "^1.1.1",
-    "destr": "^2.0.1",
+    "changelogen": "^0.5.7",
+    "config": "^3.3.12",
+    "destr": "^2.0.3",
     "escape-string-regexp": "^5.0.0",
-    "eslint": "^8.48.0",
-    "eslint-config-unjs": "^0.2.1",
-    "esm": "^3.2.25",
+    "eslint": "^9.11.0",
+    "eslint-config-unjs": "^0.3.2",
     "estree-walker": "^3.0.3",
-    "execa": "^8.0.1",
-    "fast-glob": "^3.3.1",
-    "mlly": "^1.4.2",
-    "object-hash": "^3.0.0",
-    "pathe": "^1.1.1",
-    "pirates": "^4.0.6",
-    "pkg-types": "^1.0.3",
-    "prettier": "^3.0.3",
-    "reflect-metadata": "^0.1.13",
-    "semver": "^7.5.4",
-    "std-env": "^3.4.3",
-    "terser-webpack-plugin": "^5.3.9",
-    "ts-loader": "^9.4.4",
-    "tslib": "^2.6.2",
-    "typescript": "^5.2.2",
-    "vite": "^4.4.9",
-    "vitest": "^0.34.3",
-    "webpack": "^5.88.2",
-    "webpack-cli": "^5.1.4"
+    "fast-glob": "^3.3.2",
+    "mlly": "^1.7.1",
+    "nano-jsx": "^0.1.0",
+    "pathe": "^1.1.2",
+    "pkg-types": "^1.2.0",
+    "preact": "^10.24.1",
+    "preact-render-to-string": "^6.5.11",
+    "prettier": "^3.3.3",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "reflect-metadata": "^0.2.2",
+    "solid-js": "^1.9.1",
+    "std-env": "^3.7.0",
+    "terser-webpack-plugin": "^5.3.10",
+    "tinyexec": "^0.3.0",
+    "ts-loader": "^9.5.1",
+    "typescript": "^5.6.2",
+    "vitest": "^2.1.1",
+    "vue": "^3.5.8",
+    "webpack": "^5.94.0",
+    "webpack-bundle-analyzer": "^4.10.2",
+    "webpack-cli": "^5.1.4",
+    "webpack-license-plugin": "^4.5.0",
+    "yoctocolors": "^2.1.1"
   },
-  "packageManager": "pnpm@8.7.4"
+  "packageManager": "pnpm@9.11.0"
 }
