@@ -1,16 +1,17 @@
 #!/usr/bin/env node
 
-const { resolve } = require("path");
+import { resolve } from "node:path";
+import { createJiti } from "./jiti.cjs";
 
 const script = process.argv.splice(2, 1)[0];
 
 if (!script) {
-  // eslint-disable-next-line no-console
   console.error("Usage: jiti <path> [...arguments]");
   process.exit(1);
 }
 
 const pwd = process.cwd();
-const jiti = require("..")(pwd);
+const jiti = createJiti(pwd);
 const resolved = (process.argv[1] = jiti.resolve(resolve(pwd, script)));
-jiti(resolved);
+
+jiti.import(resolved).catch(console.error);
