@@ -1,14 +1,24 @@
-import { transformSync } from "@babel/core";
 import type {
   TransformOptions as BabelTransformOptions,
   PluginItem,
 } from "@babel/core";
-import { TransformOptions, TRANSFORM_RESULT } from "./types";
+import { transformSync } from "@babel/core";
+import proposalDecoratorsPlugin from "@babel/plugin-proposal-decorators";
+import syntaxClassPropertiesPlugin from "@babel/plugin-syntax-class-properties";
+import syntaxImportAssertionsPlugin from "@babel/plugin-syntax-import-assertions";
+import transformExportNamespaceFromPlugin from "@babel/plugin-transform-export-namespace-from";
+import transformTypeScriptPlugin from "@babel/plugin-transform-typescript";
+import parameterDecoratorPlugin from "babel-plugin-parameter-decorator";
+// import transformTypeScriptMetaPlugin from "babel-plugin-transform-typescript-metadata";
+import transformTypeScriptMetaPlugin from "./plugins/babel-plugin-transform-typescript-metadata";
+import syntaxJSXPlugin from "@babel/plugin-syntax-jsx";
+import transformReactJSX from "@babel/plugin-transform-react-jsx";
 import { TransformImportMetaPlugin } from "./plugins/babel-plugin-transform-import-meta";
 import { importMetaEnvPlugin } from "./plugins/import-meta-env";
-import babelPluginTransformTypescriptMetadata from "./plugins/babel-plugin-transform-typescript-metadata";
+import transformModulesPlugin from "./plugins/transform-module";
+import type { TransformOptions, TransformResult } from "./types";
 
-export default function transform(opts: TransformOptions): TRANSFORM_RESULT {
+export default function transform(opts: TransformOptions): TransformResult {
   const _opts: BabelTransformOptions & { plugins: PluginItem[] } = {
     babelrc: false,
     configFile: false,
@@ -20,36 +30,40 @@ export default function transform(opts: TransformOptions): TRANSFORM_RESULT {
     ...opts.babel,
     plugins: [
       [
-        require("@babel/plugin-transform-modules-commonjs"),
-        { allowTopLevelThis: true },
+        transformModulesPlugin,
+        {
+          allowTopLevelThis: true,
+          noInterop: !opts.interopDefault,
+          async: opts.async,
+        },
       ],
-      [require("babel-plugin-dynamic-import-node"), { noInterop: true }],
       [TransformImportMetaPlugin, { filename: opts.filename }],
-      [require("@babel/plugin-syntax-class-properties")],
-      [require("@babel/plugin-transform-export-namespace-from")],
+      [syntaxClassPropertiesPlugin],
+      [transformExportNamespaceFromPlugin],
       [importMetaEnvPlugin],
     ],
   };
 
+  if (opts.jsx) {
+    _opts.plugins.push(
+      [syntaxJSXPlugin],
+      [transformReactJSX, Object.assign({}, opts.jsx)],
+    );
+  }
+
   if (opts.ts) {
     _opts.plugins.push([
-      require("@babel/plugin-transform-typescript"),
+      transformTypeScriptPlugin,
       { allowDeclareFields: true },
     ]);
     // `unshift` because these plugin must come before `@babel/plugin-syntax-class-properties`
     _opts.plugins.unshift(
-      [babelPluginTransformTypescriptMetadata],
-      [require("@babel/plugin-proposal-decorators"), { legacy: true }],
+      [transformTypeScriptMetaPlugin],
+      [transformTypeScriptMetaPlugin],
+      [proposalDecoratorsPlugin, { legacy: true }],
     );
-    _opts.plugins.push(require("babel-plugin-parameter-decorator"));
-    _opts.plugins.push(require("@babel/plugin-syntax-import-assertions"));
-  }
-
-  if (opts.legacy) {
-    _opts.plugins.push(
-      require("@babel/plugin-transform-nullish-coalescing-operator"),
-    );
-    _opts.plugins.push(require("@babel/plugin-transform-optional-chaining"));
+    _opts.plugins.push(parameterDecoratorPlugin);
+    _opts.plugins.push(syntaxImportAssertionsPlugin);
   }
 
   if (opts.babel && Array.isArray(opts.babel.plugins)) {
