@@ -1,10 +1,10 @@
-import { join, resolve, dirname } from "path";
-import { execa } from "execa";
+import { join, resolve, dirname } from "node:path";
+import { x } from "tinyexec";
 import { describe, it, expect } from "vitest";
 import fg from "fast-glob";
 
 describe("fixtures", async () => {
-  const jitiPath = resolve(__dirname, "../bin/jiti.js");
+  const jitiPath = resolve(__dirname, "../lib/jiti-cli.mjs");
 
   const root = dirname(__dirname);
   const dir = join(__dirname, "fixtures");
@@ -18,29 +18,43 @@ describe("fixtures", async () => {
 
       // Clean up absolute paths and sourcemap locations for stable snapshots
       function cleanUpSnap(str: string) {
-        return (str + "\n")
-          .replace(/\n\t/g, "\n")
-          .replace(/\\+/g, "/")
-          .split(cwd.replace(/\\/g, "/"))
-          .join("<cwd>") // workaround for replaceAll in Node 14
-          .split(root.replace(/\\/g, "/"))
-          .join("<root>") // workaround for replaceAll in Node 14
-          .replace(/:\d+:\d+([\s')])/g, "$1") // remove line numbers in stacktrace
-          .replace(/node:(internal|events)/g, "$1") // in Node 16 internal will be presented as node:internal
-          .replace(/\.js\)/g, ")")
-          .replace(/file:\/{3}/g, "file://")
-          .replace(/Node.js v[\d.]+/, "Node.js v<version>")
-          .replace(/ParseError: \w:\/:\s+/, "ParseError: ") // Unknown chars in Windows
-          .trim();
+        return (
+          (str + "\n")
+            .replace(/\n\t/g, "\n")
+            .replace(/\\+/g, "/")
+            .split(cwd.replace(/\\/g, "/"))
+            .join("<cwd>") // workaround for replaceAll in Node 14
+            .split(root.replace(/\\/g, "/"))
+            .join("<root>") // workaround for replaceAll in Node 14
+            .replace(/:\d+:\d+([\s')])/g, "$1") // remove line numbers in stacktrace
+            .replace(/node:(internal|events)/g, "$1") // in Node 16 internal will be presented as node:internal
+            .replace(/\.js\)/g, ")")
+            .replace(/file:\/{3}/g, "file://")
+            .replace(/Node.js v[\d.]+/, "Node.js v<version>")
+            .replace(/ParseError: \w:\/:\s+/, "ParseError: ") // Unknown chars in Windows
+            .replace("TypeError [ERR_INVALID_ARG_TYPE]:", "TypeError:")
+            .replace("eval_evalModule", "evalModule")
+            // Node 18
+            .replace(
+              "  ErrorCaptureStackTrace(err);",
+              "validateFunction(listener, 'listener');",
+            )
+            .replace("internal/errors:496", "events:276")
+            .replace("    ^", "  ^")
+            // eslint-disable-next-line no-control-regex
+            .replace(/\u001B\[[\d;]*m/gu, "")
+            .trim()
+        );
       }
 
-      const { stdout, stderr } = await execa("node", [jitiPath, fixtureEntry], {
-        cwd,
-        stdio: "pipe",
-        reject: false,
-        env: {
-          JITI_CACHE: "false",
-          JITI_ESM_RESOLVE: "true",
+      const { stdout, stderr } = await x("node", [jitiPath, fixtureEntry], {
+        nodeOptions: {
+          cwd,
+          stdio: "pipe",
+          env: {
+            JITI_CACHE: "false",
+            JITI_JSX: "true",
+          },
         },
       });
 
