@@ -1,4 +1,4 @@
 export function foo(boo: string) {
   // @ts-expect-error
-  boooooooooooo = import
+  boooooooooooo = (
 }
