@@ -1,3 +1,3 @@
 export const version = "123";
 
-export * as utils from "./utils";
+export * as utils from "./utils.mjs";
