@@ -1,15 +1,15 @@
-// eslint-disable-next-line import/no-duplicates
 import imported from "./file.json";
-// eslint-disable-next-line import/no-duplicates
+
 import importedWithAssertion from "./file.json" assert { type: "json" };
 
 const required = require("./file.json");
 
-const debug = (label: string, value) =>
+const debug = (label: string, value: any) =>
   console.log(label, ":", value, ".default:", value.default);
 
 debug("Imported", imported);
 debug("Imported with assertion", importedWithAssertion);
 debug("Required", required);
-// eslint-disable-next-line unicorn/prefer-top-level-await
-import("./file.json").then((r) => debug("Dynamic Imported", r));
+
+const dynamicRes = await import("./file.json");
+debug("Dynamic Imported", dynamicRes);
