@@ -2,10 +2,6 @@ import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   test: {
-    exclude: [
-      "**/test.{ts,mjs,cjs,js}",
-      "node_modules/**/*",
-      "test/bun.test.ts",
-    ],
+    include: ["test/fixtures.test.ts", "test/utils.test.ts"],
   },
 });
